Removing BTC fee rate updater
