+++ bootstrap/src/main/kotlin/jp/co/soramitsu/bootstrap/genesis/d3/D3TestContext.kt
@@ -65,11 +65,6 @@ object D3TestContext {
             "notary",
             listOf("withdrawal")
         ),
-        AccountPrototype(
-            "btc_fee_rate",
-            "notary",
-            listOf("btc_fee_rate_setter")
-        ),
         PeersCountDependentAccountPrototype(
             "btc_withdrawal_service",
             "notary",
@@ -349,16 +344,6 @@ object D3TestContext {
                 Primitive.RolePermission.can_grant_can_set_my_quorum,
                 Primitive.RolePermission.can_grant_can_add_my_signatory,
                 Primitive.RolePermission.can_grant_can_remove_my_signatory
-            )
-        )
-    }
-
-    fun createBtcFeeRateSetterRole(builder: TransactionBuilder) {
-        builder.createRole(
-            "btc_fee_rate_setter",
-            listOf(
-                Primitive.RolePermission.can_set_detail,
-                Primitive.RolePermission.can_get_all_accounts
             )
         )
     }
