+++ bootstrap/src/main/resources/test-genesis.block.json
@@ -227,15 +227,6 @@
               },
               {
                 "createRole": {
-                  "roleName": "btc_fee_rate_setter",
-                  "permissions": [
-                    "can_set_detail",
-                    "can_get_all_accounts"
-                  ]
-                }
-              },
-              {
-                "createRole": {
                   "roleName": "brvs",
                   "permissions": [
                     "can_add_signatory",
@@ -459,19 +450,6 @@
                   "accountName": "withdrawal",
                   "domainId": "notary",
                   "publicKey": "09274270232f43fd794c5c81762ac1c8687f7083682654656ea96d1f9b14f4d3"
-                }
-              },
-              {
-                "createAccount": {
-                  "accountName": "btc_fee_rate",
-                  "domainId": "notary",
-                  "publicKey": "1776d3cf1e73b0f361dd9287f8aa7d95f78bb6abe44784ce1cd173096084abd0"
-                }
-              },
-              {
-                "appendRole": {
-                  "accountId": "btc_fee_rate@notary",
-                  "roleName": "btc_fee_rate_setter"
                 }
               },
               {
