+++ configs/btc/withdrawal_local.properties
@@ -20,10 +20,6 @@ btc-withdrawal.btcConsensusCredential.accountId=btc_consensus_collector@notary
 btc-withdrawal.btcConsensusCredential.pubkeyPath=deploy/iroha/keys/btc_consensus_collector@notary.pub
 btc-withdrawal.btcConsensusCredential.privkeyPath=deploy/iroha/keys/btc_consensus_collector@notary.priv
 # --------- Credentials -------
-btc-withdrawal.btcFeeRateCredential.accountId=btc_fee_rate@notary
-btc-withdrawal.btcFeeRateCredential.pubkeyPath=deploy/iroha/keys/btc_fee_rate@notary.pub
-btc-withdrawal.btcFeeRateCredential.privkeyPath=deploy/iroha/keys/btc_fee_rate@notary.priv
-# --------- Credentials -------
 # TODO we don't need so much information about notary account
 btc-withdrawal.notaryCredential.accountId=notary@notary
 btc-withdrawal.notaryCredential.pubkeyPath=deploy/iroha/keys/notary0@notary.pub
