+++ configs/btc/withdrawal_mainnet.properties
@@ -20,10 +20,6 @@ btc-withdrawal.signatureCollectorCredential.accountId=btc_sign_collector@notary
 btc-withdrawal.signatureCollectorCredential.pubkeyPath=deploy/iroha/keys/btc_sign_collector@notary.pub
 btc-withdrawal.signatureCollectorCredential.privkeyPath=deploy/iroha/keys/btc_sign_collector@notary.priv
 # --------- Credentials -------
-btc-withdrawal.btcFeeRateCredential.accountId=btc_fee_rate@notary
-btc-withdrawal.btcFeeRateCredential.pubkeyPath=deploy/iroha/keys/btc_fee_rate@notary.pub
-btc-withdrawal.btcFeeRateCredential.privkeyPath=deploy/iroha/keys/btc_fee_rate@notary.priv
-# --------- Credentials -------
 btc-withdrawal.notaryCredential.accountId=notary@notary
 btc-withdrawal.notaryCredential.pubkeyPath=deploy/iroha/keys/notary0@notary.pub
 btc-withdrawal.notaryCredential.privkeyPath=deploy/iroha/keys/notary0@notary.priv
