--- deploy/iroha/keys/btc_fee_rate@notary.priv
@@ -1,1 +1,0 @@
-edb6c609a1868fa5fadea72fb218a46fb188a1d303440706593e9fbb5df6c7a3
