--- deploy/iroha/keys/btc_fee_rate@notary.pub
@@ -1,1 +1,0 @@
-1776d3cf1e73b0f361dd9287f8aa7d95f78bb6abe44784ce1cd173096084abd0
